Update GuiBio 12222019 downloadable content
This is a progressional update towards the development of Downloadable content from online gene banks. This update should show the addition of the BeautifulSoup4 webscraper being used to gather the genetic data off of a site and print it to console on the click of that specific gene's Download button. The next update will be focusing on the development of file creation/saving with file browser and deletion. Later, there will be development on file creation/saving for the comparison data.