oh boy
Added a lot of mess that needs to be cleaned before it can be useful :)  Specifically added in checkboxes on the third tab and a new function that checks to see whether the boxes have been checked or not. This is to later be used as a place where whatever you download can be seen and has a checkbox in case you want to run it in a motif search along with your added file or if you just want to do the checkboxes as a stand alone run.